[FIX] google_calendar: events disappearing reset account
Before this commit, when performing the Google Calendar account reset
in Odoo, some events and recurrences were disappearing from Odoo when
they shouldn't. This was happening because we were not removing the
'google_id' information of deleted recurrences or even unlinking
events incorrectly in Odoo side. Additionally, the synchronization of
only 'new' or 'all existing' events were not correctly being handled
because the 'need_sync' write should change according with the selected
'sync_policy'.

After this commit, we correctly remove the 'google_id' of recurrences
when the deleting them and we now consider correctly the 'sync_policy'
of synchronizing only 'new' or 'all existing' events by updating the
'need_sync' status of all existing synchronized events.

task-3731683

closes odoo/odoo#174720

X-original-commit: 1db5893eedfb4bddce1c088619bd5b806d090592
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>